# some misc microoptimizations