some minor microoptimizations that I finally cleaned up.

```
       django_template.py             4.0s (2)             4.0s (2)  -0.4%
            pyxl_bench.py             3.6s (2)             3.6s (2)  -0.2%
sqlalchemy_imperative2.py             4.8s (2)             4.8s (2)  -0.3%
                  geomean                 4.1s                 4.1s  -0.3%
```